Added `de/struct_utils` module
